dream/requirements.txt
moviepy
pillow
numpy
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
yt-dlp
